build
[git-p4: depot-paths = "//depot/qt/main/": change = 119274]

+++ src/tools/qstring.cpp
@@ -3657,7 +3657,7 @@ QString &QString::sprintf( const char* cformat, ... )
 		break;
 	    }
 	    case 'p': {
-	    	Q_ULLONG i = (Q_ULLONG)(unsigned) va_arg(ap, void*);
+	    	Q_ULLONG i = (Q_ULLONG) va_arg(ap, void*);
 		flags |= QLocalePrivate::Alternate;
 		subst = locale.d->unsLongLongToString(i, precision, 16, width, flags);
 		++c;
